Correcting bugs in script.

+++ cmps101/scripts/workspace_setup.sh
@@ -1,28 +1,41 @@
 #!/usr/bin/env bash
-cd
-workspace="cmps101-grading"
-binaries="bin"
-mkdir bin -p
-cd $binaries
-binaries=`pwd`
-cd
+if [ "$#" != "2" ] ;
+then
+    echo "USAGE: $0 setup|clean"
+    exit 1
+fi
 
-mkdir $workspace -p
-cd $workspace
-workspace=`pwd`
+if [ "$1" == "setup" ] ;
+then
 
-unlink "$binaries/dc.sh"
-unlink "$binaries/isltester"
+    cd
+    workspace="cmps101-grading"
+    binaries="bin"
+    mkdir bin -p
+    cd $binaries
+    binaries=`pwd`
+    cd
 
-wget https://github.com/alfredd/ucsc_work/archive/master.zip
-unzip master.zip  -d cmps101_scripts
+    mkdir $workspace -p
+    cd $workspace
+    workspace=`pwd`
+
+    unlink "$binaries/dc.sh"
+    unlink "$binaries/isltester"
+
+    wget https://github.com/alfredd/ucsc_work/archive/master.zip
+    unzip master.zip  -d cmps101_scripts
 
 
-cd $workspace
+    cd $workspace
 
-chmod +x "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/deadline_cloner.sh"
-ln -s "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/deadline_cloner.sh" "$binaries/dc.sh"
+    chmod +x "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/deadline_cloner.sh"
+    ln -s "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/deadline_cloner.sh" "$binaries/dc.sh"
 
 
-chmod +x "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/sorttester.py"
-ln -s "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/sorttester.py" "$binaries/isltester"
+    chmod +x "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/sorttester.py"
+    ln -s "$workspace/cmps101_scripts/ucsc_work-master/cmps101/scripts/sorttester.py" "$binaries/isltester"
+elif [ "$1" == "clean" ]
+then
+    echo "cleaning"
+fi
